// packages/protocol/typescript/generated/agent_start/runtime/v1/files_pb.d.ts
export declare type MutationResult = Message<"agent_start.runtime.v1.MutationResult"> & {
  /**
   * @generated from field: bool ok = 1;
   */
  ok: boolean;
};
// ...
export declare type FilesServiceWriteRequest = Message<"agent_start.runtime.v1.FilesServiceWriteRequest"> & {
  /**
   * @generated from field: string worktree = 1;
   */
  worktree: string;

  /**
   * @generated from field: string relative_path = 2;
   */
  relativePath: string;

  /**
   * @generated from field: string content = 3;
   */
  content: string;
};

/**
 * Describes the message agent_start.runtime.v1.FilesServiceWriteRequest.
 * Use `create(FilesServiceWriteRequestSchema)` to create a new message.
 */
export declare const FilesServiceWriteRequestSchema: GenMessage<FilesServiceWriteRequest>;

/**
 * @generated from message agent_start.runtime.v1.FilesServiceWriteResponse
 */
export declare type FilesServiceWriteResponse = Message<"agent_start.runtime.v1.FilesServiceWriteResponse"> & {
  /**
   * @generated from field: agent_start.runtime.v1.MutationResult result = 1;
   */
  result?: MutationResult | undefined;
};

/**
 * Describes the message agent_start.runtime.v1.FilesServiceWriteResponse.
 * Use `create(FilesServiceWriteResponseSchema)` to create a new message.
 */
export declare const FilesServiceWriteResponseSchema: GenMessage<FilesServiceWriteResponse>;

/**
 * @generated from message agent_start.runtime.v1.FilesServiceWriteBase64Request
 */
export declare type FilesServiceWriteBase64Request = Message<"agent_start.runtime.v1.FilesServiceWriteBase64Request"> & {
  /**
   * @generated from field: string worktree = 1;
   */
  worktree: string;

  /**
   * @generated from field: string relative_path = 2;
   */
  relativePath: string;

  /**
   * @generated from field: bytes content = 3;
   */
  content: Uint8Array;
};

/**
 * Describes the message agent_start.runtime.v1.FilesServiceWriteBase64Request.
 * Use `create(FilesServiceWriteBase64RequestSchema)` to create a new message.
 */
export declare const FilesServiceWriteBase64RequestSchema: GenMessage<FilesServiceWriteBase64Request>;

/**
 * @generated from message agent_start.runtime.v1.FilesServiceWriteBase64Response
 */
export declare type FilesServiceWriteBase64Response = Message<"agent_start.runtime.v1.FilesServiceWriteBase64Response"> & {
  /**
   * @generated from field: agent_start.runtime.v1.MutationResult result = 1;
   */
  result?: MutationResult | undefined;
};

/**
 * Describes the message agent_start.runtime.v1.FilesServiceWriteBase64Response.
 * Use `create(FilesServiceWriteBase64ResponseSchema)` to create a new message.
 */
export declare const FilesServiceWriteBase64ResponseSchema: GenMessage<FilesServiceWriteBase64Response>;

/**
 * @generated from message agent_start.runtime.v1.FilesServiceWriteBase64ChunkRequest
 */
export declare type FilesServiceWriteBase64ChunkRequest = Message<"agent_start.runtime.v1.FilesServiceWriteBase64ChunkRequest"> & {
  /**
   * @generated from field: string worktree = 1;
   */
  worktree: string;

  /**
   * @generated from field: string relative_path = 2;
   */
  relativePath: string;

  /**
   * @generated from field: bytes content = 3;
   */
  content: Uint8Array;

  /**
   * @generated from field: bool append = 4;
   */
  append: boolean;
};

/**
 * Describes the message agent_start.runtime.v1.FilesServiceWriteBase64ChunkRequest.
 * Use `create(FilesServiceWriteBase64ChunkRequestSchema)` to create a new message.
 */
export declare const FilesServiceWriteBase64ChunkRequestSchema: GenMessage<FilesServiceWriteBase64ChunkRequest>;

/**
 * @generated from message agent_start.runtime.v1.FilesServiceWriteBase64ChunkResponse
 */
export declare type FilesServiceWriteBase64ChunkResponse = Message<"agent_start.runtime.v1.FilesServiceWriteBase64ChunkResponse"> & {
  /**
   * @generated from field: agent_start.runtime.v1.MutationResult result = 1;
   */
  result?: MutationResult | undefined;
};

/**
 * Describes the message agent_start.runtime.v1.FilesServiceWriteBase64ChunkResponse.
 * Use `create(FilesServiceWriteBase64ChunkResponseSchema)` to create a new message.
 */
export declare const FilesServiceWriteBase64ChunkResponseSchema: GenMessage<FilesServiceWriteBase64ChunkResponse>;

/**
 * @generated from message agent_start.runtime.v1.FilesServiceCreateFileRequest
 */
export declare type FilesServiceCreateFileRequest = Message<"agent_start.runtime.v1.FilesServiceCreateFileRequest"> & {
  /**
   * @generated from field: string worktree = 1;
   */
  worktree: string;

  /**
   * @generated from field: string relative_path = 2;
   */
  relativePath: string;
};

/**
 * Describes the message agent_start.runtime.v1.FilesServiceCreateFileRequest.
 * Use `create(FilesServiceCreateFileRequestSchema)` to create a new message.
 */
export declare const FilesServiceCreateFileRequestSchema: GenMessage<FilesServiceCreateFileRequest>;

/**
 * @generated from message agent_start.runtime.v1.FilesServiceCreateFileResponse
 */
export declare type FilesServiceCreateFileResponse = Message<"agent_start.runtime.v1.FilesServiceCreateFileResponse"> & {
  /**
   * @generated from field: agent_start.runtime.v1.MutationResult result = 1;
   */
  result?: MutationResult | undefined;
};

/**
 * Describes the message agent_start.runtime.v1.FilesServiceCreateFileResponse.
 * Use `create(FilesServiceCreateFileResponseSchema)` to create a new message.
 */
export declare const FilesServiceCreateFileResponseSchema: GenMessage<FilesServiceCreateFileResponse>;

/**
 * @generated from message agent_start.runtime.v1.FilesServiceCreateDirectoryRequest
 */
export declare type FilesServiceCreateDirectoryRequest = Message<"agent_start.runtime.v1.FilesServiceCreateDirectoryRequest"> & {
  /**
   * @generated from field: string worktree = 1;
   */
  worktree: string;

  /**
   * @generated from field: string relative_path = 2;
   */
  relativePath: string;
};

/**
 * Describes the message agent_start.runtime.v1.FilesServiceCreateDirectoryRequest.
 * Use `create(FilesServiceCreateDirectoryRequestSchema)` to create a new message.
 */
export declare const FilesServiceCreateDirectoryRequestSchema: GenMessage<FilesServiceCreateDirectoryRequest>;

/**
 * @generated from message agent_start.runtime.v1.FilesServiceCreateDirectoryResponse
 */
export declare type FilesServiceCreateDirectoryResponse = Message<"agent_start.runtime.v1.FilesServiceCreateDirectoryResponse"> & {
  /**
   * @generated from field: agent_start.runtime.v1.MutationResult result = 1;
   */
  result?: MutationResult | undefined;
};

/**
 * Describes the message agent_start.runtime.v1.FilesServiceCreateDirectoryResponse.
 * Use `create(FilesServiceCreateDirectoryResponseSchema)` to create a new message.
 */
export declare const FilesServiceCreateDirectoryResponseSchema: GenMessage<FilesServiceCreateDirectoryResponse>;

/**
 * @generated from message agent_start.runtime.v1.FilesServiceCreateDirectoryNoClobberRequest
 */
export declare type FilesServiceCreateDirectoryNoClobberRequest = Message<"agent_start.runtime.v1.FilesServiceCreateDirectoryNoClobberRequest"> & {
  /**
   * @generated from field: string worktree = 1;
   */
  worktree: string;

  /**
   * @generated from field: string relative_path = 2;
   */
  relativePath: string;
};

/**
 * Describes the message agent_start.runtime.v1.FilesServiceCreateDirectoryNoClobberRequest.
 * Use `create(FilesServiceCreateDirectoryNoClobberRequestSchema)` to create a new message.
 */
export declare const FilesServiceCreateDirectoryNoClobberRequestSchema: GenMessage<FilesServiceCreateDirectoryNoClobberRequest>;

/**
 * @generated from message agent_start.runtime.v1.FilesServiceCreateDirectoryNoClobberResponse
 */
export declare type FilesServiceCreateDirectoryNoClobberResponse = Message<"agent_start.runtime.v1.FilesServiceCreateDirectoryNoClobberResponse"> & {
  /**
   * @generated from field: agent_start.runtime.v1.MutationResult result = 1;
   */
  result?: MutationResult | undefined;
};

/**
 * Describes the message agent_start.runtime.v1.FilesServiceCreateDirectoryNoClobberResponse.
 * Use `create(FilesServiceCreateDirectoryNoClobberResponseSchema)` to create a new message.
 */
export declare const FilesServiceCreateDirectoryNoClobberResponseSchema: GenMessage<FilesServiceCreateDirectoryNoClobberResponse>;

/**
 * @generated from message agent_start.runtime.v1.FilesServiceCommitUploadRequest
 */
export declare type FilesServiceCommitUploadRequest = Message<"agent_start.runtime.v1.FilesServiceCommitUploadRequest"> & {
  /**
   * @generated from field: string worktree = 1;
   */
  worktree: string;

  /**
   * @generated from field: string temp_relative_path = 2;
   */
  tempRelativePath: string;

  /**
   * @generated from field: string final_relative_path = 3;
   */
  finalRelativePath: string;
};

/**
 * Describes the message agent_start.runtime.v1.FilesServiceCommitUploadRequest.
 * Use `create(FilesServiceCommitUploadRequestSchema)` to create a new message.
 */
export declare const FilesServiceCommitUploadRequestSchema: GenMessage<FilesServiceCommitUploadRequest>;

/**
 * @generated from message agent_start.runtime.v1.FilesServiceCommitUploadResponse
 */
export declare type FilesServiceCommitUploadResponse = Message<"agent_start.runtime.v1.FilesServiceCommitUploadResponse"> & {
  /**
   * @generated from field: agent_start.runtime.v1.MutationResult result = 1;
   */
  result?: MutationResult | undefined;
};

/**
 * Describes the message agent_start.runtime.v1.FilesServiceCommitUploadResponse.
 * Use `create(FilesServiceCommitUploadResponseSchema)` to create a new message.
 */
export declare const FilesServiceCommitUploadResponseSchema: GenMessage<FilesServiceCommitUploadResponse>;

/**
 * @generated from message agent_start.runtime.v1.FilesServiceRenameRequest
 */
export declare type FilesServiceRenameRequest = Message<"agent_start.runtime.v1.FilesServiceRenameRequest"> & {
  /**
   * @generated from field: string worktree = 1;
   */
  worktree: string;

  /**
   * @generated from field: string old_relative_path = 2;
   */
  oldRelativePath: string;

  /**
   * @generated from field: string new_relative_path = 3;
   */
  newRelativePath: string;
};

/**
 * Describes the message agent_start.runtime.v1.FilesServiceRenameRequest.
 * Use `create(FilesServiceRenameRequestSchema)` to create a new message.
 */
export declare const FilesServiceRenameRequestSchema: GenMessage<FilesServiceRenameRequest>;

/**
 * @generated from message agent_start.runtime.v1.FilesServiceRenameResponse
 */
export declare type FilesServiceRenameResponse = Message<"agent_start.runtime.v1.FilesServiceRenameResponse"> & {
  /**
   * @generated from field: agent_start.runtime.v1.MutationResult result = 1;
   */
  result?: MutationResult | undefined;
};

/**
 * Describes the message agent_start.runtime.v1.FilesServiceRenameResponse.
 * Use `create(FilesServiceRenameResponseSchema)` to create a new message.
 */
export declare const FilesServiceRenameResponseSchema: GenMessage<FilesServiceRenameResponse>;

/**
 * @generated from message agent_start.runtime.v1.FilesServiceCopyRequest
 */
export declare type FilesServiceCopyRequest = Message<"agent_start.runtime.v1.FilesServiceCopyRequest"> & {
  /**
   * @generated from field: string worktree = 1;
   */
  worktree: string;

  /**
   * @generated from field: string source_relative_path = 2;
   */
  sourceRelativePath: string;

  /**
   * @generated from field: string destination_relative_path = 3;
   */
  destinationRelativePath: string;
};

/**
 * Describes the message agent_start.runtime.v1.FilesServiceCopyRequest.
 * Use `create(FilesServiceCopyRequestSchema)` to create a new message.
 */
export declare const FilesServiceCopyRequestSchema: GenMessage<FilesServiceCopyRequest>;

/**
 * @generated from message agent_start.runtime.v1.FilesServiceCopyResponse
 */
export declare type FilesServiceCopyResponse = Message<"agent_start.runtime.v1.FilesServiceCopyResponse"> & {
  /**
   * @generated from field: agent_start.runtime.v1.MutationResult result = 1;
   */
  result?: MutationResult | undefined;
};

/**
 * Describes the message agent_start.runtime.v1.FilesServiceCopyResponse.
 * Use `create(FilesServiceCopyResponseSchema)` to create a new message.
 */
export declare const FilesServiceCopyResponseSchema: GenMessage<FilesServiceCopyResponse>;

/**
 * @generated from message agent_start.runtime.v1.FilesServiceDeleteRequest
 */
export declare type FilesServiceDeleteRequest = Message<"agent_start.runtime.v1.FilesServiceDeleteRequest"> & {
  /**
   * @generated from field: string worktree = 1;
   */
  worktree: string;

  /**
   * @generated from field: string relative_path = 2;
   */
  relativePath: string;

  /**
   * @generated from field: bool recursive = 3;
   */
  recursive: boolean;
};

/**
 * Describes the message agent_start.runtime.v1.FilesServiceDeleteRequest.
 * Use `create(FilesServiceDeleteRequestSchema)` to create a new message.
 */
export declare const FilesServiceDeleteRequestSchema: GenMessage<FilesServiceDeleteRequest>;

/**
 * @generated from message agent_start.runtime.v1.FilesServiceDeleteResponse
 */
export declare type FilesServiceDeleteResponse = Message<"agent_start.runtime.v1.FilesServiceDeleteResponse"> & {
  /**
   * @generated from field: agent_start.runtime.v1.MutationResult result = 1;
   */
  result?: MutationResult | undefined;
};

/**
 * Describes the message agent_start.runtime.v1.FilesServiceDeleteResponse.
 * Use `create(FilesServiceDeleteResponseSchema)` to create a new message.
 */
export declare const FilesServiceDeleteResponseSchema: GenMessage<FilesServiceDeleteResponse>;